move scf to extern
cuz don’t want to test
