+++ discO/extern/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+# see LICENSE.rst
+
+"""External codes, untested and unloved."""
+
+
+##############################################################################
+# END
